Create UI for multiplayer
Added NetworkServer.Reset() in the Disconnect method, otherwise it sends the error "StartServer listen failed".
The button StartHost start a game with you being the host, the button JoinGame allows you to join a game after gaving the ip address of the host.
NetworkManager is moved to the lobby scene.